requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.0.1
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
django-filter==23.4
django-extensions==3.2.3